requirements.txt
# requirements.txt

alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
bcrypt==3.2.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.3
MarkupSafe==2.1.5
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
orjson==3.10.0
pillow==10.2.0
pycparser==2.21
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
PyMySQL==1.1.0
python-barcode==0.15.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.37.2
sympy==1.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
zipp==3.17.0
